Update
- Added licensing and a gitignore file
- Converted AuthState in PlayerData to an enumeration
- Added SetStatistic to MobTemplate and fixed broken unit tests
- Completed At, AtObj, Advance, and Authorize commands

+++ Source/SmaugCS/act_move.cs
@@ -331,7 +331,7 @@ namespace SmaugCS
                     return ReturnTypes.None;
             }
 
-            var dtxt = rev_exit((int)xit.Direction);
+            var dtxt = rev_exit(xit.Direction);
 
             // First determine if the player should be moved or not Check various flags, spells, 
             // the players position and strength vs. the pull, etc... any kind of checks you like.
